+++ ExilePartyClient/src/app/shared/providers/message-value.service.ts
@@ -7,8 +7,10 @@ import { RobotService } from './robot.service';
 @Injectable()
 export class MessageValueService {
   public playerGain = 0;
+  public currentPlayerGain = 0;
   public partyGain = 0;
   public playerValue = 0;
+  public currentPlayerValue = 0;
   public partyValue = 0;
 
   private defaultMsg = '%No data registered';
@@ -35,7 +37,7 @@ export class MessageValueService {
 
   updateMessages() {
     // tslint:disable-next-line:max-line-length
-    this.playerNetworthMsg = `[ExileParty] My net worth: ${this.playerValue.toFixed(1)}c. Gain: ${this.playerGain.toFixed(1)}c / hour`;
+    this.playerNetworthMsg = `[ExileParty] My net worth: ${this.currentPlayerValue.toFixed(1)}c. Gain: ${this.currentPlayerGain.toFixed(1)}c / hour`;
     // tslint:disable-next-line:max-line-length
     this.partyNetworthMsg = `[ExileParty] Grp net worth: ${this.partyValue.toFixed(1)}c. Gain: ${this.partyGain.toFixed(1)}c / hour`;
   }